# Simplify BufferedReader to improve DataLoader performance