--- a/python/paddle/fluid/tests/unittests/test_decoupled_py_reader.py
+++ b/python/paddle/fluid/tests/unittests/test_decoupled_py_reader.py
@@ -122,8 +122,14 @@ class TestBase(unittest.TestCase):
                             label = item['label']
                             assert image.shape() == [BATCH_SIZE, 784]
                             assert label.shape() == [BATCH_SIZE, 1]
-                            assert image._place()._equals(ps[i])
-                            assert label._place()._equals(ps[i])
+                            if ps[i]._equals(fluid.CPUPlace()):
+                                assert image._place()._equals(fluid.CPUPlace())
+                                assert label._place()._equals(fluid.CPUPlace())
+                            else:
+                                assert image._place()._equals(
+                                    fluid.CUDAPinnedPlace())
+                                assert label._place()._equals(
+                                    fluid.CUDAPinnedPlace())
                         L, = exe.run(program=prog,
                                      feed=d,
                                      fetch_list=[loss],